Move `@ExperimentalRestartApi` to package
